src/scraparr/requirements.txt
prometheus_client
requests
werkzeug
python-dateutil
pyyaml
python-dotenv